+++ packages/terra-avatar/tests/wdio/avatar-common-spec.js
@@ -25,8 +25,14 @@ Terra.describeViewports('Avatar', ['huge'], () => {
     Terra.it.matchesScreenshot({ selector: '#image-avatar' });
   });
 
-  describe('Invalid Image Avatar', () => {
-    before(() => browser.url('/#/raw/tests/terra-avatar/avatar/avatar/invalid-image-avatar'));
+  describe('Invalid Image Avatar With Initials', () => {
+    before(() => browser.url('/#/raw/tests/terra-avatar/avatar/avatar/invalid-image-avatar-with-initials'));
+
+    Terra.it.validatesElement({ selector: '#invalid-image-avatar' });
+  });
+
+  describe('Invalid Image Avatar Without Initials', () => {
+    before(() => browser.url('/#/raw/tests/terra-avatar/avatar/avatar/invalid-image-avatar-without-initials'));
 
     Terra.it.validatesElement({ selector: '#invalid-image-avatar' });
   });